+++ tests/placement_hints/test_ca1.py
@@ -18,7 +18,7 @@ class Test_ca1(unittest.TestCase):
             padding=10, layer_thickness=30, x_thickness=35, z_thickness=30
         )
         direction_vectors = np.zeros(
-            cls.ca1_mock.annotation.raw.shape + (3,), dtype=np.float
+            cls.ca1_mock.annotation.raw.shape + (3,), dtype=float
         )
         direction_vectors[cls.ca1_mock.annotation.raw > 0] = (0.0, -1.0, 0.0)
         cls.result = tested.compute_placement_hints(
